requirements.txt
timm
streamlit  
pillow  
numpy
keras
matplotlib
tensorflow
torch
torchvision
os
